# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Rucaro Accounting API
  version: 1.0.0
  summary: 日本語会計アプリ Rucaro の REST API（新基盤）
  description: |
    Rucaro Accounting 新基盤（PHP 8.3 + MariaDB 10.11）の REST API 仕様。

    対象スコープ:
      - 認証（opaque Bearer token, DB 保管、SHA-256 ハッシュ化）
      - 会計主体（entity）・会計期（fiscal term）管理
      - 勘定科目（account title）・補助科目（sub account title）マスタ参照
      - 仕訳（journal entry）の CRUD と検索（借方・貸方・税率・摘要）
      - 試算表（trial balance）の集計出力
      - 決算書（financial statements: BS / PL / CS）の出力
      - 固定資産（fixed asset）台帳と減価償却バッチ
      - 領収書（receipt）アップロード → Claude 抽出 → 承認 → 仕訳化のパイプライン
      - メール / メッセージ承認トークン処理
      - ヘルスチェック

    共通レスポンス包絡形式:
    ```json
    { "success": true, "data": { ... }, "error": null, "meta": null }
    ```

    エラー時:
    ```json
    { "success": false, "data": null, "error": { "code": "VALIDATION_FAILED", "message": "...", "details": {} }, "meta": null }
    ```

    ID は全て ULID (Crockford Base32, 26 文字)。金額は文字列 DECIMAL(18,4) で送受信。
    日付は ISO 8601（暦日は `YYYY-MM-DD`、時刻込みは RFC 3339 UTC、例 `2026-04-21T12:34:56.789012Z`）。
    通貨は ISO 4217 3 文字コード。既定 `JPY`。

    借方合計 = 貸方合計 の不変条件は OpenAPI レベルでは記述のみ、実装側（Application 層）で強制する。
  termsOfService: https://example.com/terms
  contact:
    name: Rucaro Accounting Maintainers
    url: https://github.com/masudagreen/accounting
    email: masudagreen@gmail.com
  license:
    name: GPL-2.0-or-later
    identifier: GPL-2.0-or-later
...
servers:
  - url: http://localhost:8080/api/v1
    description: ローカル開発環境（Docker Compose）
...
security:
  - bearerAuth: []
paths:
...
  /approvals/{token}:
    parameters:
      - name: token
        in: path
        required: true
        description: ワンタイム承認トークン（URL 中の不透明文字列）。SHA-256 ハッシュ化して DB 参照する。
        schema:
          type: string
          minLength: 32
          maxLength: 128
          pattern: "^[A-Za-z0-9_-]{32,128}$"
    post:
      tags: [Approval]
      operationId: respondApproval
      summary: 承認 / 却下
      description: |
        メール / メッセージで送信したリンクから承認または却下を受け付ける。
        承認時は対象仕訳を `approved` + `posted` に、却下時は `rejected` に遷移。
        トークンは 1 度だけ使用可能。期限切れ・使用済みは 410 を返す。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalResponse"
      responses:
        "200":
          description: 承認 / 却下確定
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ApprovalResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          description: トークン期限切れ / 使用済み
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        opaque token を `Authorization: Bearer <token>` ヘッダで送信。
        サーバは SHA-256 ハッシュで `api_tokens.token_hash` と突き合わせる。
...
  responses:
    BadRequest:
      description: リクエストが不正（JSON 破損・必須パラメータ欠如・型違い）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
...
    NotFound:
      description: 指定 ID のリソースが存在しない、または論理削除済み
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
...
    UnprocessableEntity:
      description: スキーマは通ったがドメインルール違反（借貸不一致 等）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
    ServerError:
      description: サーバ内部エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
  schemas:
...
    Ulid:
      type: string
      description: Crockford Base32 の 26 文字 ULID
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
      minLength: 26
      maxLength: 26
      example: 01HW7K9B2QV7C8Y4ZEXAMPLE00
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 マイクロ秒精度。UTC 送受信、表示時に Asia/Tokyo へ変換
      example: "2026-04-21T03:15:27.123456Z"
...
    Error:
      type: object
      description: エラー情報
      properties:
        code:
          type: string
          description: 機械可読エラーコード（UPPER_SNAKE_CASE）
          example: VALIDATION_FAILED
        message:
          type: string
          description: 人間可読メッセージ（日本語可）
          example: 借方合計と貸方合計が一致していません
        details:
          type: object
          description: フィールド別エラーなど追加詳細
          additionalProperties: true
          nullable: true
      required: [code, message]
    Meta:
      type: object
      description: ページネーションメタ情報
      properties:
        total:
          type: integer
          minimum: 0
          example: 128
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          example: 50
      required: [total, page, pageSize]
    Envelope:
      type: object
      description: 成功レスポンスの包絡
      properties:
        success:
          type: boolean
          example: true
        data:
          description: 各エンドポイント固有のデータ。失敗時は null。
          nullable: true
        error:
          oneOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
        meta:
          oneOf:
            - $ref: "#/components/schemas/Meta"
            - type: "null"
      required: [success, data, error]
    EnvelopeError:
      type: object
      description: エラーレスポンスの包絡
      properties:
        success:
          type: boolean
          const: false
        data:
          type: "null"
        error:
          $ref: "#/components/schemas/Error"
        meta:
          oneOf:
            - $ref: "#/components/schemas/Meta"
            - type: "null"
      required: [success, data, error]

    # -----------------------------------------------------------
    # ヘルス
    # -----------------------------------------------------------
...
    ApprovalDecision:
      type: string
      enum: [approved, rejected]
    ApprovalChannel:
      type: string
      enum: [email, line, slack, discord]
    ApprovalResponse:
      type: object
      properties:
        decision:
          $ref: "#/components/schemas/ApprovalDecision"
        comment:
          type: string
          maxLength: 512
          default: ""
      required: [decision]
    ApprovalResult:
      type: object
      properties:
        tokenId:
          $ref: "#/components/schemas/Ulid"
        decision:
          $ref: "#/components/schemas/ApprovalDecision"
        respondedAt:
          $ref: "#/components/schemas/Timestamp"
        channel:
          $ref: "#/components/schemas/ApprovalChannel"
        journalEntryId:
          oneOf:
            - $ref: "#/components/schemas/Ulid"
            - type: "null"
        receiptId:
          oneOf:
            - $ref: "#/components/schemas/Ulid"
            - type: "null"
      required: [tokenId, decision, respondedAt, channel]